Rearrange translation code
